+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Beautified system emails
 - Added admin theme font subset latin, latin-ext
+- Fixed simple plugin regex
 
 ### 0.8.4 (July 25, 2014)